+++ lib/providers/games.dart
@@ -2,12 +2,14 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:poll_and_play/config.dart';
 import 'package:poll_and_play/grpc/games.dart';
+import 'package:poll_and_play/providers/provider.dart';
 import 'package:poll_play_proto_gen/public.dart';
 
-class GamesProvider extends ChangeNotifier {
+class GamesProvider extends ChangeNotifier implements Provider {
   final GamesClient _client = GamesClient(GlobalConfig().apiAddress.split(':'));
   late List<GameWithStat> _games;
 
+  @override
   Future<void> init() async {
     if (FirebaseAuth.instance.currentUser == null) {
       return;